gui/requirements.txt
PyGObject>=3.42.0
pyyaml>=6.0
